Allow to configure timeouts via system properties / configuration files
### Configuration

Test framework allows to customise the configuration for running the test case via a `test.properties` file placed under `src/test/resources` folder.

All the properties can be configured globally by replacing `<YOUR SERVICE NAME>` with `global`.

- Timeouts

Timeouts are quite important property to, as an example, control how long to wait for a service to start. At the moment, 
the only available timeout is startup of services:
 
```
ts.<YOUR SERVICE NAME>.startup.timeout=10m
```

For example, we can use `ts.pingApp.startup.timeout` to wait up to 10 min for the ping application to start. 

How often does the test framework check for startup conditions? This is also configurable via:

```
ts.<YOUR SERVICE NAME>.startup.check-poll-interval=2s
```

Now, we're saying that we need to wait up to 10 minutes for our service to start and will be checking the condition every 2 seconds.

How can we manipulate the overall timeout in slower environments? Let's say that our environment is twice and a half slower than a good environment, then we can instruct the test framework with:

```
ts.global.factor.timeout=2.5
```

And all the timeouts will take into account this factor value. For example, if previously, the startup timeout was 10 minutes, now it will be 10 * 2.5 = 25 minutes.