Support S2i for Native builds
In order to manage the base image in use, you need to provide the properties:
- For JVM: `ts.global.s2i.quarkus.jvm.builder.image`
- For Native: `ts.global.s2i.quarkus.native.builder.image`

The way these properties are up to users. In the examples, we supply this configuration in the pom.xml as part of system properties (in the Maven failsafe plugin). 
But we can provide a custom property by service in the `test.properties` file. For further information about how to customise the properties, go to the Configuration section.