Change port resolution strategy from random to incremental
The previous port resolution was based on Spring implementation and it was random. The problem is that we were hitting in port conflicts sometimes. 

This commit addresses the situation by changing the strategy to find an available port in an incremental manner. This implementation is based on the Apache Camel implementation.