Refactor solution for Metrics ++ Polish properties
The changes in this PR are:
- Change the tracing/metrics format to follow the same format as the rest of the properties

This change allows to enable/disable/configure these extensions per module.

- Refactor the metrics module to decouple the metrics of prometheus and have another type of exporter: file.

My idea is to integrate the metrics within the running Quarkus application, so we can assert metrics during the test execution without needing to have a Prometheus.

Also, we're exporting the metrics into a file (/target/logs/metrics.out).