Adapt quarkus test core to bootstrap setForcedDependencies change
The method setForcedDependencies signature changed from `List<AppDependency>` to `List<Dependency>` where Dependency is an interface of AppDependency, so it should be fine. However, the compiler fails to cast it, so we need to use reflection to sort it out for the most recent version and older versions.

This change in upstream was added as part of this commit: https://github.com/quarkusio/quarkus/commit/657ae2965955ba1bd015d2049e4830dc223c1b36