Add support for a separate managed inteface
Ability to have a management interface on a separate port (and host) was
added[1] to Quarkus upstream.
These changes are required to properly cover this feature in TS.
Noticeable user-facing feature is addition of `management` method to `RestService` class
This method either returns a way to communicate to the interface or
fails, if separate interface is not enabled.

[1] https://github.com/quarkusio/quarkus/pull/30506