Print application logs only after start-up
Previously[1] it was (wrongly) assumed, that method isRunning of QuarkusManagedResource class
is run only once after the start of the application, so it is safe to put log output into it.
In reality, this method is called several times during test lifecycle,
including at the time of teardown.
Additionally, logs for OpenShiftQuarkusApplicationManagedResource are retrieved from
OCP cluster. Since connection to the cluster may be already closed
during teardown, this leads to hard-to-reproduce exceptions.
The change addresses this problem by making sure, that logs are printed only
once during start up and not in (supposedly) pure isRunning method.

[1] https://github.com/quarkus-qe/quarkus-test-framework/pull/213