Fix nightly failures
Before that change we used quarkus CLI properties as environment
variables in openshift manifests (which we should not have do) and
for some reason it worked (which it should not have do).
When this behaviour was changed upstream, our jobs started to fail,
this change should fix that.

See https://github.com/quarkusio/quarkus/issues/36359 for details