Add propagation of maven.repo.local on windows machine
In case of linux we propagating all build arguments which also contains `maven.repo.local`.
This is not case on windows where we getting only specific set of them (starts with Quarkus).
When #1359 was introducedThe build start separatly using maven.
But when provided path to local maven repo was set as cmd argument and the artifact is not public,
causing the windows failing the build of app with that forced artifact.

(cherry picked from commit afdcab9fc63cf7f9f0e40326e3bee5c50b813aa2)