Change the Keycloak default realm file name
The name is modified to follow Keycloak convention.
The correct naming should be `<realm-name>-realm.json` to import the realm from specific directory
or the name can be anything which don't ending with reaml.json to load it all config from one file.

This is based on discussion on https://github.com/keycloak/keycloak/issues/36284#issuecomment-2577748182