Fix check release version GH action for older releases
In case ofstable releases the script check if the initial release was released
the GitHub give the response contains the default value of entries (30).
The initial release was not found in the response list so it failed the release check.
This PR raising the number or returned releases from 30 to 100 items.